♻️ Move test for ErrorHandler from sqlslog package to opts package
